# ProyectoFinalAvanzada/src/main/gen/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API para reportar incidentes ciudadanos en tiempo real.
  title: Plataforma de Reportes de Seguridad
  version: 1.0.0
servers:
- description: Servidor de producción
  url: https://api.reportes-ciudad.com/v1
- description: Servidor local para desarrollo
  url: http://localhost:8080
paths:
  /auth/register:
    post:
      description: Registra un nuevo usuario en la plataforma.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRegistro'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_register_post_201_response'
          description: Usuario registrado con éxito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_register_post_400_response'
          description: Error en la solicitud
      summary: Registro de usuario
      tags:
      - user
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: user
  /auth/verify:
    post:
      description: Permite a un usuario verificar su cuenta con el código enviado
        por correo electrónico.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/verificarCodigo'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_verify_post_200_response'
          description: Cuenta verificada con éxito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_verify_post_400_response'
          description: Código incorrecto o expirado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_verify_post_404_response'
          description: Usuario no encontrado
      summary: Verificación de cuenta
      tags:
      - user
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: user
  /auth/login:
    post:
      description: Permite a un usuario autenticarse en la plataforma.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_login_post_200_response'
          description: Inicio de sesión exitoso
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_login_post_401_response'
          description: Credenciales inválidas
      summary: Inicio de sesión
      tags:
      - user
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: user
      - tag: admin
  /auth/perfil:
    delete:
      description: Permite a un usuario eliminar su cuenta de la plataforma.
      parameters:
      - description: Debe enviarse como 'true' para confirmar la eliminación.
        explode: true
        in: query
        name: confirmar
        required: true
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_perfil_delete_200_response'
          description: Perfil eliminado con éxito.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_perfil_delete_400_response'
          description: Falta confirmación para eliminar la cuenta.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_perfil_delete_401_response'
          description: No autorizado.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_perfil_delete_500_response'
          description: Error interno del servidor.
      summary: Eliminar perfil de usuario
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
      - tag: admin
    put:
      description: "Permite al usuario modificar su información personal, incluyendo\
        \ nombre, ubicación, correo y contraseña."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioEditar'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_perfil_put_200_response'
          description: Datos actualizados correctamente.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_perfil_put_400_response'
          description: Error en la solicitud
      summary: Editar perfil del usuario
      tags:
      - user
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: user
...
components:
  schemas:
    UsuarioRegistro:
      example:
        ubicacion:
          latitud: 0.8008281904610115
          longitud: 6.027456183070403
        ciudadResidencia: ciudadResidencia
        correo: correo
        direccion: direccion
        id: id
        telefono: telefono
        nombre: nombre
        contraseña: contraseña
      properties:
        id:
          type: string
        nombre:
          type: string
        ciudadResidencia:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        correo:
          format: email
          type: string
        contraseña:
          format: password
          type: string
        ubicacion:
          $ref: '#/components/schemas/UsuarioRegistro_ubicacion'
      required:
      - ciudadResidencia
      - contraseña
      - correo
      - direccion
      - id
      - nombre
      - telefono
      - ubicacion
      type: object
    verificarCodigo:
      properties:
        correo:
          example: usuario@email.com
          format: email
          type: string
        codigo:
          example: A1B2C3
          type: string
      required:
      - codigo
      - correo
      type: object
    UsuarioEditar:
      example:
        ubicacion:
          latitud: 0.8008281904610115
          longitud: 6.027456183070403
        ciudadResidencia: ciudadResidencia
        correo: correo
        direccion: direccion
        telefono: telefono
        nombre: nombre
        contraseña: contraseña
      properties:
        nombre:
          type: string
        ciudadResidencia:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        correo:
          format: email
          type: string
        contraseña:
          format: password
          type: string
        ubicacion:
          $ref: '#/components/schemas/UsuarioRegistro_ubicacion'
      required:
      - correo
      type: object
    Login:
      example:
        correo: correo
        contraseña: contraseña
      properties:
        correo:
          format: email
          type: string
        contraseña:
          format: password
          type: string
      required:
      - contraseña
      - correo
      type: object
...
    _auth_register_post_201_response:
      example:
        mensaje: Usuario registrado con éxito. Verifique su correo electrónico.
      properties:
        mensaje:
          example: Usuario registrado con éxito. Verifique su correo electrónico.
          type: string
      type: object
    _auth_register_post_400_response:
      example:
        error: El correo ya está registrado.
      properties:
        error:
          example: El correo ya está registrado.
          type: string
      type: object
    _auth_verify_post_200_response:
      example:
        mensaje: Cuenta activada con éxito.
      properties:
        mensaje:
          example: Cuenta activada con éxito.
          type: string
      type: object
    _auth_verify_post_400_response:
      example:
        error: Código inválido o expirado.
      properties:
        error:
          example: Código inválido o expirado.
          type: string
      type: object
    _auth_verify_post_404_response:
      example:
        error: Usuario no encontrado.
      properties:
        error:
          example: Usuario no encontrado.
          type: string
      type: object
    _auth_login_post_200_response_usuario:
      example:
        id: id
        nombre: nombre
        rol: rol
      properties:
        id:
          type: string
        nombre:
          type: string
        rol:
          type: string
      type: object
    _auth_login_post_200_response:
      example:
        usuario:
          id: id
          nombre: nombre
          rol: rol
        token: token
      properties:
        token:
          type: string
        usuario:
          $ref: '#/components/schemas/_auth_login_post_200_response_usuario'
      type: object
    _auth_login_post_401_response:
      example:
        error: Correo o contraseña incorrectos.
      properties:
        error:
          example: Correo o contraseña incorrectos.
          type: string
      type: object
    _auth_perfil_put_200_response:
      example:
        mensaje: Datos actualizados correctamente.
      properties:
        mensaje:
          example: Datos actualizados correctamente.
          type: string
      type: object
    _auth_perfil_put_400_response:
      example:
        error: No se puede actualizar el perfil con los datos proporcionados.
      properties:
        error:
          example: No se puede actualizar el perfil con los datos proporcionados.
          type: string
      type: object
    _auth_perfil_delete_200_response:
      example:
        mensaje: Cuenta eliminada correctamente.
      properties:
        mensaje:
          example: Cuenta eliminada correctamente.
          type: string
      type: object
    _auth_perfil_delete_400_response:
      example:
        error: Debes confirmar la eliminación de tu cuenta.
      properties:
        error:
          example: Debes confirmar la eliminación de tu cuenta.
          type: string
      type: object
    _auth_perfil_delete_401_response:
      example:
        error: No tienes permiso para eliminar esta cuenta.
      properties:
        error:
          example: No tienes permiso para eliminar esta cuenta.
          type: string
      type: object
    _auth_perfil_delete_500_response:
      example:
        error: Ocurrió un error inesperado. Inténtelo más tarde.
      properties:
        error:
          example: Ocurrió un error inesperado. Inténtelo más tarde.
          type: string
      type: object
...
    UsuarioRegistro_ubicacion:
      example:
        latitud: 0.8008281904610115
        longitud: 6.027456183070403
      properties:
        latitud:
          format: double
          type: number
        longitud:
          format: double
          type: number
      type: object